requirements.txt
asgiref==3.6.0
certifi==2022.12.7
channels==4.0.0
charset-normalizer==3.1.0
Django==4.1.7
django-jazzmin==2.6.0
django-tinymce==3.6.1
idna==3.4
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2023.2
urllib3==1.26.15
